+++ Makefile
@@ -5,12 +5,11 @@ TARGET = main
 all:$(TARGET)
 
 $(TARGET): ./src/$(TARGET).c
-	$(CC) $(CFLAGS) -o $(TARGET) ./src/$(TARGET).c ./src/utils.c ./src/currency.c ./src/unit.c ./src/mass.c ./src/length.c ./src/temp.c
-
+	$(CC) $(CFLAGS) -o $(TARGET) ./src/$(TARGET).c ./src/utils.c
 uninstall:
 	rm -f $(TARGET)
 
 clean:
 	rm -f $(TARGET)
-	$(CC) $(CFLAGS) -o $(TARGET) ./src/$(TARGET).c ./src/utils.c ./src/currency.c ./src/unit.c ./src/mass.c ./src/length.c ./src/temp.c
+	$(CC) $(CFLAGS) -o $(TARGET) ./src/$(TARGET).c ./src/utils.c
